==> 03_1 RandomForest 랜덤포레스트 학습.ipynb <==
from google.colab import drive
drive.mount('/gdrive')

from sklearn.model_selection import train_test_split
from sklearn.model_selection import ParameterGrid
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier as RFC
from sklearn import metrics
from sklearn.metrics import f1_score
from sklearn.model_selection import ParameterGrid
import numpy as np
import pandas as pd
from pprint import pprint
import warnings
warnings.filterwarnings('ignore')


# from kaggler.preprocessing import LabelEncoder
# from kaggler.model import AutoLGB

# 1. 전처리 된 파일 불러오기
- 텍스트데이터 제거
- 밀린것들 정리
- 파이썬으로 작업

meta2017 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2017_pre.csv')
meta2018 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2018_pre.csv')
meta2019 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2019_pre.csv')
meta2020 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2020_pre.csv')

# 2. 연도별 학습
- 콜랩에선 램 한게때문에 로컬로 진행

## 2-1. 2017년

X = meta2017.drop(['knowcode' ], axis = 1)
y = meta2017['knowcode']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=30)

학습해 볼 여러 파라미터를 설정해줌

# 2017년 
param_grid = dict() 
param_grid_for_rfc = ParameterGrid({"max_depth": range(20,40,5),
                                    "n_estimators": range(200,375,25),
                                    "random_state" : [10,20,30]})


param_grid[RFC] = param_grid_for_rfc

best_score = -1 #초기값 설정

for model_func in [RFC]:
    
    for param in param_grid[model_func]:
        model = model_func(**param).fit(train_x, train_y)
        pred_Y = model.predict(test_x)        
        score = f1_score(test_y, pred_Y, average = 'macro')
        print("모델:", model_func)
        print("점수:", score)
        print("파라미터:", param)
        
        if score > best_score: 
            # 현재 점수가 지금까지 찾은 최고 점수보다 좋으면, 최고 모델, 파라미터, 점수 업데이트
            best_model_func = model_func
            best_score = score
            best_param = param
        print("모델:", best_model_func)
        print("점수:", best_score)
        print("파라미터:", best_param)
        print()

# 현재 학습된 파라미터의 점수와 최고치를 보여준 파라미터의 점수를 같이 보여줌

## 2-1. 2018년

X = meta2018.drop(['knowcode' ], axis = 1)
y = meta2018['knowcode']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=30)

학습해 볼 여러 파라미터를 설정해줌

# 2017년 
param_grid = dict() 
param_grid_for_rfc = ParameterGrid({"max_depth": range(20,40,5),
                                    "n_estimators": range(200,375,25),
                                    "random_state" : [10,20,30]})


param_grid[RFC] = param_grid_for_rfc

best_score = -1 #초기값 설정

for model_func in [RFC]:
    
    for param in param_grid[model_func]:
        model = model_func(**param).fit(train_x, train_y)
        pred_Y = model.predict(test_x)        
        score = f1_score(test_y, pred_Y, average = 'macro')
        print("모델:", model_func)
        print("점수:", score)
        print("파라미터:", param)
        
        if score > best_score: 
            # 현재 점수가 지금까지 찾은 최고 점수보다 좋으면, 최고 모델, 파라미터, 점수 업데이트
            best_model_func = model_func
            best_score = score
            best_param = param
        print("모델:", best_model_func)
        print("점수:", best_score)
        print("파라미터:", best_param)
        print()

# 현재 학습된 파라미터의 점수와 최고치를 보여준 파라미터의 점수를 같이 보여줌

## 2-3. 2019년

X = meta2019.drop(['knowcode' ], axis = 1)
y = meta2019['knowcode']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=30)

학습해 볼 여러 파라미터를 설정해줌

# 2017년 
param_grid = dict() 
param_grid_for_rfc = ParameterGrid({"max_depth": range(20,40,5),
                                    "n_estimators": range(200,375,25),
                                    "random_state" : [10,20,30]})


param_grid[RFC] = param_grid_for_rfc

best_score = -1 #초기값 설정

for model_func in [RFC]:
    
    for param in param_grid[model_func]:
        model = model_func(**param).fit(train_x, train_y)
        pred_Y = model.predict(test_x)        
        score = f1_score(test_y, pred_Y, average = 'macro')
        print("모델:", model_func)
        print("점수:", score)
        print("파라미터:", param)
        
        if score > best_score: 
            # 현재 점수가 지금까지 찾은 최고 점수보다 좋으면, 최고 모델, 파라미터, 점수 업데이트
            best_model_func = model_func
            best_score = score
            best_param = param
        print("모델:", best_model_func)
        print("점수:", best_score)
        print("파라미터:", best_param)
        print()

# 현재 학습된 파라미터의 점수와 최고치를 보여준 파라미터의 점수를 같이 보여줌

## 2-4. 2020년

X = meta2020.drop(['knowcode' ], axis = 1)
y = meta2020['knowcode']

X_train, X_test, y_train, y_test = train_test_split(X, y, stratify=y, random_state=30)

학습해 볼 여러 파라미터를 설정해줌

# 2017년 
param_grid = dict() 
param_grid_for_rfc = ParameterGrid({"max_depth": range(20,40,5),
                                    "n_estimators": range(200,375,25),
                                    "random_state" : [10,20,30]})


param_grid[RFC] = param_grid_for_rfc

best_score = -1 #초기값 설정

for model_func in [RFC]:
    
    for param in param_grid[model_func]:
        model = model_func(**param).fit(train_x, train_y)
        pred_Y = model.predict(test_x)        
        score = f1_score(test_y, pred_Y, average = 'macro')
        print("모델:", model_func)
        print("점수:", score)
        print("파라미터:", param)
        
        if score > best_score: 
            # 현재 점수가 지금까지 찾은 최고 점수보다 좋으면, 최고 모델, 파라미터, 점수 업데이트
            best_model_func = model_func
            best_score = score
            best_param = param
        print("모델:", best_model_func)
        print("점수:", best_score)
        print("파라미터:", best_param)
        print()

# 현재 학습된 파라미터의 점수와 최고치를 보여준 파라미터의 점수를 같이 보여줌

# 3. Bayesian Optimization을 통한 파라미터 찾기

!pip install bayesian-optimization

from bayes_opt import BayesianOptimization

def rf_bo(max_depth, n_estimators, random_state):

  rf_params = {
              'max_depth' : int(round(max_depth)),
              'n_estimators' : int(round(n_estimators)), 
              'random_state' : int(round(random_state)
              }

  rf = RFC(**rf_params)
  X_train, X_valid, y_train, y_valid = train_test_split(X,y,test_size = 0.2, stratify = y)

  rf.fit(X_train,y_train)
  score = f1_score(y_valid, rf.predict(X_valid), average = 'macro')

  return score

## 3.1 2017년

학습범위설정

rf_parameter_bounds = {
                      'max_depth' : (20,35),
                      'n_estimators' : (175,350),   
                      'random_state' : (10,35)
                      }

BO_rf_2017 = BayesianOptimization(f = rf_bo, pbounds = rf_parameter_bounds)

학습

하이퍼 파라미터 숫자들을 더한만큼 학습됨

BO_rf_2017.maximize(init_points = 5, n_iter = 30)

## 3.1 2018년

학습범위설정

rf_parameter_bounds = {
                      'max_depth' : (20,35),
                      'n_estimators' : (175,350),   
                      'random_state' : (10,35)
                      }

BO_rf_2018 = BayesianOptimization(f = rf_bo, pbounds = rf_parameter_bounds)

학습

하이퍼 파라미터 숫자들을 더한만큼 학습됨

BO_rf_2018.maximize(init_points = 5, n_iter = 30)

## 3.1 2019년

학습범위설정

rf_parameter_bounds = {
                      'max_depth' : (20,35),
                      'n_estimators' : (175,350),   
                      'random_state' : (10,35)
                      }

BO_rf_2019 = BayesianOptimization(f = rf_bo, pbounds = rf_parameter_bounds)

학습

하이퍼 파라미터 숫자들을 더한만큼 학습됨

BO_rf_2019.maximize(init_points = 5, n_iter = 30)

## 3.1 2020년

학습범위설정

rf_parameter_bounds = {
                      'max_depth' : (20,35),
                      'n_estimators' : (175,350),   
                      'random_state' : (10,35)
                      }

BO_rf_2020 = BayesianOptimization(f = rf_bo, pbounds = rf_parameter_bounds)

학습

하이퍼 파라미터 숫자들을 더한만큼 학습됨

BO_rf_2020.maximize(init_points = 5, n_iter = 30)
