==> 03_3 voting Classifier로 수치데이터 학습 통합 .ipynb <==
from google.colab import drive
drive.mount('/gdrive')

import pandas as pd
import numpy as np
import warnings
warnings.filterwarnings('ignore')
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import ExtraTreesClassifier
from sklearn.ensemble import VotingClassifier

# 1. 전처리 된 파일 불러오기
- 텍스트데이터 제거
- 밀린것들 정리
- 파이썬으로 작업

학습용

meta2017 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2017_pre.csv')
meta2018 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2018_pre.csv')
meta2019 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2019_pre.csv')
meta2020 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2020_pre.csv')

테스트용

test2017 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2017_pre_test.csv')
test2018 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2018_pre_test.csv')
test2019 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2019_pre_test.csv')
test2020 = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Data_preprocessed/2차 전처리 : 파이썬 작업/KNOW_2020_pre_test.csv')

# 2. 연도별로 학습을 통해 찾은 파라미터를 적용해 KNOWCODE를 테스트에 추가한다
- Voting Classifier로 두가지 모델의 성능을 통합한 제출 파일을 만든다

2017년

et = ExtraTreesClassifier(max_depth = 30, n_estimators = 300, random_state = 30)
rf = RandomForestClassifier(max_depth = 25, n_estimators = 350)

model = VotingClassifier(estimators=[('rf',rf),('et',et)],voting = 'soft')

train_x = meta2017.drop(['knowcode'], axis = 1)
train_y = meta2017['knowcode']
test_x = test2017

model.fit(train_x, train_y)
pred2017 = model.predict(test_x)

2018년

et = ExtraTreesClassifier(max_depth = 30, n_estimators = 300, random_state = 30)
rf = RandomForestClassifier(max_depth = 25, n_estimators = 350)

model = VotingClassifier(estimators=[('rf',rf),('et',et)],voting = 'soft')

train_x = meta2018.drop(['knowcode'], axis = 1)
train_y = meta2018['knowcode']
test_x = test2018

model.fit(train_x, train_y)
pred2018 = model.predict(test_x)

2019년

et = ExtraTreesClassifier(max_depth = 30, n_estimators = 300, random_state = 30)
rf = RandomForestClassifier(max_depth = 25, n_estimators = 350)

model = VotingClassifier(estimators=[('rf',rf),('et',et)],voting = 'soft')

train_x = meta2019.drop(['knowcode'], axis = 1)
train_y = meta2019['knowcode']
test_x = test2019

model.fit(train_x, train_y)
pred2019 = model.predict(test_x)

2020년

et = ExtraTreesClassifier(max_depth = 30, n_estimators = 300, random_state = 30)
rf = RandomForestClassifier(max_depth = 25, n_estimators = 350)

model = VotingClassifier(estimators=[('rf',rf),('et',et)],voting = 'soft')

train_x = meta2020.drop(['knowcode'], axis = 1)
train_y = meta2020['knowcode']
test_x = test2020

model.fit(train_x, train_y)
pred2020 = model.predict(test_x)

# 3. 위에서 만든것들 통합하기

sub = pd.read_csv('/gdrive/MyDrive/2조/data/코드정리/Submission/sample_submission.csv')

sub

통합

pred_y = pred2017.tolist() + pred2018.tolist() + pred2019.tolist() + pred2020.tolist()

sub['knowcode'] = pred_y

sub

sub.to_csv('/gdrive/MyDrive/2조/data/코드정리/Submission/sub_1차.csv', index = False)
